Adds support for non-Pacific timezones
This is the simplest approach--adds all valid US time zones (as
determined by Rails) to the dropdown when creating a district. Then
displays the timezone along with the time on all public-facing pages.

Google Calendar should support without any changes.

Would ideally set this off of the geocode, since the list is already
kind of overwhelming even though it's not comprehensive. I'm also not
sure how well it will deal with applying timezones from the string in
the database (e.g. for Arizona).