Refactor ordering of elected officials & meetings
Sets a default sort order on the police district model for elected
officials, so that we don't have to declare it in each controller when
passing to the view.

I noticed the sort order in our admin section was a little weird,
because we hadn't declared it in the admin controller. This should fix
that.